rpi-server/requirements.txt
# Core dependencies
pyserial==3.5
paho-mqtt==2.2.1
python-dotenv==1.0.0
rpi-lora==0.2.1  # For LoRa communication
pyLoRaRF==1.1.3  # Alternative LoRa library
loguru==0.7.0    # Better logging
